newscaster/tts/requirements.txt
pika
google-cloud-texttospeech
google-cloud-storage
python-dotenv
pydub
redis
prometheus_client
psycopg2-binary